port ClientTimeoutSpec to munit
